Add client_id to analytics events
